# docs/mkdocs/api/swagger.yaml
basePath: /
definitions:
...
  api.addMemberRequest:
    properties:
      address:
        type: string
      label:
        type: string
    type: object
...
  api.dashboardJSON:
    properties:
      chain_id:
        type: string
      confirmed_ready:
        type: integer
      final_genesis_sha256:
        type: string
      genesis_time:
        type: string
      launch_id:
        type: string
      status:
        type: string
      threshold_status:
        type: string
      total_approved:
        type: integer
      validators:
        items:
          $ref: '#/definitions/api.validatorReadinessJSON'
        type: array
      voting_power_confirmed:
        type: number
    type: object
  api.errorBody:
    properties:
      code:
        type: string
      invariants:
        items:
          $ref: '#/definitions/api.invariantResultJSON'
        type: array
      message:
        type: string
      request_id:
        type: string
    type: object
  api.errorEnvelope:
    properties:
      error:
        $ref: '#/definitions/api.errorBody'
    type: object
...
  api.invariantResultJSON:
    properties:
      invariant:
        type: string
      ok:
        type: boolean
      reason:
        type: string
    type: object
...
  api.memberJSON:
    properties:
      added_at:
        type: string
      added_by:
        type: string
      address:
        type: string
      label:
        type: string
    type: object
...
  api.peerJSON:
    properties:
      operator_address:
        type: string
      peer_address:
        type: string
    type: object
  api.peersResponse:
    properties:
      peers:
        items:
          $ref: '#/definitions/api.peerJSON'
        type: array
    type: object
...
  api.validatorReadinessJSON:
    properties:
      binary_hash_confirmed:
        type: string
      genesis_hash_confirmed:
        type: string
      is_ready:
        type: boolean
      join_request_id:
        type: string
      last_confirmed_at:
        type: string
      moniker:
        type: string
      operator_address:
        type: string
      voting_power_pct:
        type: number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Chain launch coordination server. Coordinators and validators authenticate
    via secp256k1 challenge-response (ADR-036)
  title: ChainCoord API
  version: "1.0"
paths:
...
  /launch/{id}/dashboard:
    get:
      description: Returns combined launch and readiness status for all approved validators.
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.dashboardJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      summary: Get launch dashboard
      tags:
      - readiness
...
  /launch/{id}/members:
    get:
      description: Returns the launch's members list (address + label + provenance),
        sorted by address. Committee members only.
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.memberJSON'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      security:
      - BearerAuth: []
      summary: List launch members
      tags:
      - members
    post:
      consumes:
      - application/json
      description: |-
        Adds a hot actor address (with a label) to the launch's members list — granting see + submit access.
        Committee members only. Allowed only in DRAFT/PUBLISHED/WINDOW_OPEN. Idempotent on address.
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      - description: Address and label to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.addMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.memberJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Add a launch member
      tags:
      - members
  /launch/{id}/members/{address}:
    delete:
      description: |-
        Removes a hot actor address from the launch's members list — revoking see + submit access.
        Committee members only. Allowed only in DRAFT/PUBLISHED/WINDOW_OPEN.
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      - description: Hot actor address to remove
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Remove a launch member
      tags:
      - members
...
  /launch/{id}/peers:
    get:
      description: Returns peer addresses of all approved validators. Use ?format=text
        for persistent_peers format (comma-separated).
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      - description: Output format
        enum:
        - json
        - text
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JSON format; with ?format=text returns comma-separated text/plain
          schema:
            $ref: '#/definitions/api.peersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      summary: Get approved validator peers
      tags:
      - readiness
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Session token obtained from POST /auth/verify. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"